+++ src/main/java/net/jobsaddon/JobsAddonMain.java
@@ -7,6 +7,7 @@ import net.fabricmc.api.ModInitializer;
 import net.jobsaddon.init.CommandInit;
 import net.jobsaddon.init.ConfigInit;
 import net.jobsaddon.init.CriteriaInit;
+import net.jobsaddon.init.EventInit;
 import net.jobsaddon.init.JsonReaderInit;
 import net.jobsaddon.init.TagInit;
 import net.jobsaddon.network.JobsServerPacket;
@@ -23,6 +24,7 @@ public class JobsAddonMain implements ModInitializer {
         CriteriaInit.init();
         JobsServerPacket.init();
         TagInit.init();
+        EventInit.init();
     }
 
 }