+++ src/main/java/net/jobsaddon/mixin/compat/JobsAddonMixinPlugin.java
@@ -26,8 +26,6 @@ public class JobsAddonMixinPlugin implements IMixinConfigPlugin {
             return false;
         if (mixinClassName.contains("BlockWrapperMixin") && !FabricLoader.getInstance().isModLoaded("fallingtree"))
             return false;
-        if (mixinClassName.contains("ChopResultMixin") && !FabricLoader.getInstance().isModLoaded("treechop"))
-            return false;
         if ((mixinClassName.contains("ScytheItemMixin") || mixinClassName.contains("MacheteItemMixin")) && !FabricLoader.getInstance().isModLoaded("harvest_scythes"))
             return false;